Sorted functions
According to uncle Bob, it is nicer if the fuctions that are being called come
next. It is more difficult to read the code it the called funcion is last, as
you might have to scroll back up to continue reading the code.

The goal is to read only by scrolling down, almost like reading prose.

Also fixed some tiny PEP8.